per stream stats (#326)
* Passed cuda_stream_id from accelsim to gpgpusim

* Fixed for (auto k : kernels_info) in main() for multi-stream app

* Passed streamID of last finished kernel to gpgpu_sim::print_stats

* checkout correct gpgpu-sim branch for multi stream stats

* Move Jenkins to bak

* Adding merge_group trigger to worklow

* rename jenkinsfile

* Automated clang-format

* update formatter

* correct format script path

---------

+++ gpu-simulator/trace-parser/trace_parser.h
@@ -88,7 +88,7 @@ struct kernel_trace_t {
   unsigned tb_dim_z;
   unsigned shmem;
   unsigned nregs;
-  unsigned long cuda_stream_id;
+  unsigned long long cuda_stream_id;
   unsigned binary_verion;
   unsigned enable_lineinfo;
   unsigned trace_verion;
